preparazione-esame/django/exercise_2-3_like/2-3_like1/app/serializers.py
from rest_framework import serializers

from app.models import Book


class BookSerializer(serializers.ModelSerializer):
    class Meta:
        fields = ('id', 'title', 'author', 'genre', 'rating')
        model = Book
